// apps/git-client/src/generated/FileContent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type FileContent = { "kind": "text", path: string, content: string, sizeBytes: number, lineCount: number, } | { "kind": "binary", path: string, sizeBytes: number, } | { "kind": "invalidUtf8", path: string, sizeBytes: number, } | { "kind": "tooLarge", path: string, sizeBytes: number, lineCount: number | null, } | { "kind": "missing", path: string, };